requirements.txt
selenium==4.18.1
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0